+++ clientside/package.json
@@ -19,6 +19,7 @@
     "clsx": "^2.1.0",
     "date-fns": "^3.6.0",
     "framer-motion": "^11.0.14",
+    "js-cookie": "^3.0.5",
     "lucide-react": "^0.358.0",
     "moment": "^2.30.1",
     "next": "13.5.5",
@@ -30,6 +31,7 @@
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
+    "@types/js-cookie": "^3.0.6",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
